ci: workflow to build oops-mesa; downloader compile fix
- .github/workflows/mesa.yml: on-demand build of upstream Mesa for the target
  (pinned FreeBSD source + Mesa submodule, Docker toolchain, cached), publishing
  the archives the USE_MESA titles need. Isolated (workflow_dispatch, nothing
  depends on it) so it can't regress the working release while it stabilises.
- oops-app-downloader: cast surf->width to int (-Wsign-conversion).

+++ src/oops-utilities/oops-app-downloader/oops-app-downloader.c
@@ -46,7 +46,7 @@ int oops_dl_render(oops_surface_t *surf, const oops_dl_status_t *st) {
     if (!surf || !surf->pixels || !st) return 0;
 
     oops_draw_clear(surf, BG);
-    oops_draw_rect(surf, 0, 0, surf->width, 6, ACCENT);
+    oops_draw_rect(surf, 0, 0, (int)surf->width, 6, ACCENT);
 
     int rows = 0;
     int y = 70;
